Added OSC connection & dispatch classes
Added OSCDispatch - initial pass at allowing code to register calbacks for OSC message addresses.

Added OSCConnection, a convernience wrapper for a IBackChannelConnection to simplify the sending, receipt, and dispatch of OSC messages and bundles. Messages are received on a background thread and dispatched when requested.

Misc header and interface cleanup